src/requirements.txt
django==4.2
djangorestframework==3.15.2
django-environ==0.11.2
django-axes==8.0.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
defusedxml==0.8.0rc2
deprecation==2.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.7
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg2_binary==2.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
social-auth-core==4.5.4
sqlparse==0.5.2
tomli==2.2.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
zope.interface==7.2
celery==5.4.0
channels==3.0.4
channels-redis==3.3.1
daphne
celery-progress
channels_redis
django-cors-headers==4.6.0
aiofiles==24.1.0
flower==2.0.1
drf_yasg==1.21.8 
djoser
gunicorn==21.2.0

celery[redis]
django-redis==5.4.0